requirements.txt
einops==0.8.0
eval==0.0.5
ftfy==6.2.3
matplotlib==3.9.1
numpy==1.24.4
Pillow==10.4.0
python-dotenv==1.0.1
regex==2023.12.25
scipy==1.14.0
setuptools==69.1.1
SQLAlchemy==2.0.32
tensorboardX==2.6.2.2
timm==1.0.8
torch==2.2.1
torchvision==0.17.1
tqdm==4.66.2
opencv-python-headless==4.8.1.78
pandas
pyodbc
deltalake